make tab header part of app header
did this to stick tabheader to top on scroll
