+++ lib/main.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:url_strategy/url_strategy.dart';
-import 'package:visibility_detector/visibility_detector.dart';
-import './header.dart';
-import './tab_body.dart';
+import './home.dart';
 
 void main() {
   setPathUrlStrategy();
@@ -19,40 +17,5 @@ class MyApp extends StatelessWidget {
           primarySwatch: Colors.lightBlue,
         ),
         home: SafeArea(child: HomeScreen()));
-  }
-}
-
-class HomeScreen extends StatefulWidget {
-  @override
-  _HomeScreenState createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  bool _isMobileShowcaseVisible = true;
-
-  @override
-  Widget build(BuildContext context) {
-    bool _isMobile = MediaQuery.of(context).size.width < 600;
-
-    return Scaffold(
-        appBar: (_isMobile && _isMobileShowcaseVisible) || !_isMobile
-            ? PreferredSize(
-                preferredSize: Size.fromHeight(288),
-                child: Header(),
-              )
-            : null,
-        body: ListView(children: [
-          if (_isMobile)
-            VisibilityDetector(
-              key: Key('my-widget-key'),
-              onVisibilityChanged: (vis) {
-                setState(() {
-                  _isMobileShowcaseVisible = vis.visibleFraction == 0;
-                });
-              },
-              child: MobileShowcase(),
-            ),
-          TabBody()
-        ]));
   }
 }
